BUG Don't change date when changing account
Don't reset the date selector when the user changes the account selection. Previously, if the user selected a date with the date picker, the date would reset when changing accounts. This is generally undesired. We need account names when determining the maximum possible date range, so use `State` to bring in the account field without updating when it changes.