REF Simplify simulated portfolio creation
The new method is slightly different methodologically, but gives very similar results. Instead of converting action by action, assume that all contributions and withdrawals go to / come from the simulated security directly. This is much faster than the previous method, and sets up the option to easily simulate individual portfolios or even portions of a portfolio.