ENH Counterfactual portfolios in plots
Users now have the option to compare a displayed portfolio to a counterfactual portfolio in which all purchases and withdrawals are made in a specified security. Only update the simulated portfolio on a button click -- otherwise we'll try to download quotes for lots of invalid symbols and probably exceed the API limit for downloading quotes.