REF Refactor as a Flask app
Refactor so that the Dash app is embedded in a Flask app I control, following the example at https://medium.com/@olegkomarov_77860/how-to-embed-a-dash-app-into-an-existing-flask-app-ea05d7a2210b (code at https://github.com/okomarov/dash_on_flask). This will give me more control over the environment, and, in particular, let me use flask-sqlalchemy to set up a database for storage instead of relying on CSVs.