MAINT Update for AlphaVantage API changes
The daily adjusted timeseries endpoint is now "premium". We're only using that endpoint for dividend information, and it's not really important to have the dividend information down to the day. Therefore, we can use the standard "DAILY" endpoint and merge dividend information on a weekly basis, using two API calls.